Move write_image() to a separate source file
